Move VideoExporter into it's own source/header I'll need som help to get this building on Mac & Linux
